Patched for newer .NET version

+++ RSHelpers/PSARCUtil.cs
@@ -310,10 +310,10 @@ namespace RockSnifferLib.RSHelpers
                                     // discernable difference between notes
                                     //noteDataHash.AppendData(BitConverter.GetBytes(note.Hash));
 
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.StringIndex));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.FretId));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.AnchorFretId));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.AnchorWidth));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.StringIndex));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.FretId));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.AnchorFretId));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.AnchorWidth));
 
                                     // Ignore note.ChordId because it can differ even when there is
                                     // no discernable difference between notes
@@ -343,13 +343,13 @@ namespace RockSnifferLib.RSHelpers
                                     // Ignore note.ParentPrevNote (maybe modified by CDLC repair, maybe ignored altogether?)
                                     //noteDataHash.AppendData(BitConverter.GetBytes(note.ParentPrevNote));
 
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.SlideTo));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.SlideUnpitchTo));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.LeftHand));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.Tap));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.PickDirection));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.Slap));
-                                    noteDataHash.AppendData(BitConverter.GetBytes(note.Pluck));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.SlideTo));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.SlideUnpitchTo));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.LeftHand));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.Tap));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.PickDirection));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.Slap));
+                                    noteDataHash.AppendData(BitConverter.GetBytes((short)note.Pluck));
                                     noteDataHash.AppendData(BitConverter.GetBytes(note.Vibrato));
                                     noteDataHash.AppendData(BitConverter.GetBytes((float)Math.Round(note.Sustain, 3)));
                                     noteDataHash.AppendData(BitConverter.GetBytes((float)Math.Round(note.MaxBend, 3)));
@@ -359,8 +359,8 @@ namespace RockSnifferLib.RSHelpers
                                         noteDataHash.AppendData(BitConverter.GetBytes((float)Math.Round(bendData.Time, 3)));
                                         noteDataHash.AppendData(BitConverter.GetBytes((float)Math.Round(bendData.Step, 3)));
                                         noteDataHash.AppendData(BitConverter.GetBytes(bendData.Unk3_0));
-                                        noteDataHash.AppendData(BitConverter.GetBytes(bendData.Unk4_0));
-                                        noteDataHash.AppendData(BitConverter.GetBytes(bendData.Unk5));
+                                        noteDataHash.AppendData(BitConverter.GetBytes((short)bendData.Unk4_0));
+                                        noteDataHash.AppendData(BitConverter.GetBytes((short)bendData.Unk5));
                                     }
 
                                     if (note.FretId == 255)
@@ -381,8 +381,8 @@ namespace RockSnifferLib.RSHelpers
                                                     noteDataHash.AppendData(BitConverter.GetBytes((float)Math.Round(bendData32.Time, 3)));
                                                     noteDataHash.AppendData(BitConverter.GetBytes((float)Math.Round(bendData32.Step, 3)));
                                                     noteDataHash.AppendData(BitConverter.GetBytes(bendData32.Unk3_0));
-                                                    noteDataHash.AppendData(BitConverter.GetBytes(bendData32.Unk4_0));
-                                                    noteDataHash.AppendData(BitConverter.GetBytes(bendData32.Unk5));
+                                                    noteDataHash.AppendData(BitConverter.GetBytes((short)bendData32.Unk4_0));
+                                                    noteDataHash.AppendData(BitConverter.GetBytes((short)bendData32.Unk5));
                                                 }
 
                                                 noteDataHash.AppendData(BitConverter.GetBytes(bendData.UsedCount));
@@ -390,12 +390,12 @@ namespace RockSnifferLib.RSHelpers
 
                                             foreach (var slideTo in chordNotes.SlideTo)
                                             {
-                                                noteDataHash.AppendData(BitConverter.GetBytes(slideTo));
+                                                noteDataHash.AppendData(BitConverter.GetBytes((short)slideTo));
                                             }
 
                                             foreach (var slideUnpitchTo in chordNotes.SlideUnpitchTo)
                                             {
-                                                noteDataHash.AppendData(BitConverter.GetBytes(slideUnpitchTo));
+                                                noteDataHash.AppendData(BitConverter.GetBytes((short)slideUnpitchTo));
                                             }
 
                                             foreach (var vibrato in chordNotes.Vibrato)
